projects/notion-reporter-20260306/requirements.txt
notion-client>=2.0.0,<3.0.0
python-dotenv>=1.0.0
